Breast cancer Detection Code
The code is designed to detect breast cancer using machine learning techniques. This may  training a model on a dataset of mammography images, with the goal of classifying whether or not a given image contains evidence of breast cancer. The model may be trained using a variety of techniques, such as decision trees, support vector machines, or neural networks. Once trained, the model can then be used to predict the presence of breast cancer in new, unseen mammography images. This type of breast cancer detection can be helpful in identifying breast cancer at an early stage, when it is most treatable. It can also help to reduce the number of unnecessary biopsies by accurately identifying benign breast abnormalities.